% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_Matrix.R
\name{alias.glm2Matrix}
\alias{alias.glm2Matrix}
\title{Final Aliases in a Model using Matrix-based QR decomposition}
\usage{
\method{alias}{glm2Matrix}(object, complete = TRUE, partial = FALSE, partial.pattern = FALSE, ...)
}
\arguments{
\item{object}{A fitted GLM object with Matrix-based QR decomposition, for example one fit using \code{glm2(..., method = "glm.fit2.Matrix")}.}

\item{complete}{Should information on complete aliasing be included?}

\item{partial}{Should information on partial aliasing be included?}

\item{partial.pattern}{Should partial aliasing be presented in a schematic way? If this is done, the results are presented in a more compact way, usually giving the deciles of the coefficients.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list containing aliasing information, including coefficients expressing dependent
  columns as linear combinations of independent ones. See documentation for \code{\link[stats]{alias}} for full details.
}
\description{
This function identifies complete aliasing (perfect linear dependencies) among predictors
in a GLM model using a QR decomposition from the Matrix package. It mimics the logic of
base R's \code{alias()} but works with sparse or dense Matrix objects.
}
\details{
See documentation for \code{\link[stats]{alias}} for full details.

\code{complete = TRUE} requires the package \CRANpkg{MASS} (recommended by \CRANpkg{stats}) to be installed.
}
\note{
Although some parts of this function employ Matrix-based functions, densification (\code{as.matrix}) is unavoidable and can be slow or memory intensive for large models.
}
\examples{

data(heart)

start.p <- sum(heart$Deaths)/sum(heart$Patients)

# Model with a linearly dependent column
fit1 <- glm2(cbind(Deaths,Patients-Deaths) ~ factor(AgeGroup) + factor(Severity) +
 I(AgeGroup > 1), data = heart, family = binomial(link="log"), 
 start = c(log(start.p), rep(0, 5)), method = "glm.fit2.Matrix", x = TRUE)

alias(fit1, partial = TRUE, partial.pattern = TRUE)

}
\seealso{
\code{\link[stats]{alias}}
}
\author{
Uses the code from \code{\link[stats]{alias}}, whose authors are
listed in its documentation. Modifications to this code were made by Mark W. Donoghoe.
}
