% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glm2.R
\name{.vcov.aliased.Matrix}
\alias{.vcov.aliased.Matrix}
\title{Augment a vcov - matrix by NA rows & cols when needed:}
\usage{
.vcov.aliased.Matrix(aliased, vc, complete = TRUE)
}
\description{
Augment a vcov - matrix by NA rows & cols when needed:
}
