% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm2.influence.R
\name{dfbeta.glm2Matrix}
\alias{dfbeta.glm2Matrix}
\alias{hatvalues.glm2Matrix}
\alias{influence.measures}
\alias{dfbetas.glm2Matrix}
\title{Regression Deletion Diagnostics (partially implemented)}
\usage{
\method{dfbeta}{glm2Matrix}(model, infl = lm.influence.glm2Matrix(model, do.coef = TRUE), ...)

\method{hatvalues}{glm2Matrix}(model, infl = lm.influence.glm2Matrix(model, do.coef = FALSE), ...)

\method{influence}{measures}(model, infl = influence(model))

\method{dfbetas}{glm2Matrix}(model, infl = lm.influence.glm2Matrix(model, do.coef = TRUE), ...)
}
\value{
See documentation for \code{\link[stats]{influence.measures}}.
}
\description{
A suite of functions computing regression (leave-one-out deletion)
diagnostics for GLMs fit using \code{glm2(..., method = "\link{glm.fit2.Matrix}")}. 
Currently only \code{dfbetas} and \code{dfbeta} are implemented.
}
\seealso{
\code{\link[stats]{influence.measures}}.
}
