% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_Matrix.R
\name{kappa.glm2Matrix}
\alias{kappa.glm2Matrix}
\title{Condition Number for glm2Matrix Fits}
\usage{
\method{kappa}{glm2Matrix}(z, ...)
}
\arguments{
\item{z}{A fitted \code{glm2Matrix} object.}

\item{...}{Additional arguments passed to internal methods.
See \code{\link[base]{kappa}} for details, and the Details below for
a note about the \code{exact} argument.}
}
\value{
The condition number, \eqn{\kappa}, or an approximation if \code{exact = FALSE}.
}
\description{
Computes the condition number of the design matrix for a
\code{glm2Matrix} fit (a \code{glm} object fit using 
\code{method = "glm.fit2Matrix"}).
}
\details{
This method is based on the QR decomposition used internally to fit
the model.

It calls \code{\link[base]{kappa}} on the dense R matrix, and therefore
does not provide any efficiency benefits for large design matrices over
a model fit using \code{\link[stats]{glm}} (or \code{glm2} without \code{Matrix}).

Additionally, when \code{exact = FALSE} (the default), the returned value 
may not agree with that returned by \code{\link[base]{kappa.lm}} applied to
an equivalent dense fit. This is because the sparse QR decomposition may
have employed column pivoting so that the R matrix corresponds to a 
column-permuted design matrix. This can be checked by inspecting
\code{z$qr_full$qr@q}.
}
\seealso{
\code{\link[base]{kappa}}, \code{\link[Matrix]{sparseQR-class}}
}
