% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm2.influence.R
\name{influence.glm2Matrix}
\alias{influence.glm2Matrix}
\title{Regression Diagnostics (not yet implemented)}
\value{
An error indicating the method is not implemented.
}
\description{
This method is a placeholder. The \code{\link[stats]{influence}}
method for GLMs fit using \code{glm2(..., method = "\link{glm.fit2.Matrix}")} 
is not yet implemented.
}
