% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.glm2.R
\name{predict.glm2Matrix}
\alias{predict.glm2Matrix}
\title{Predict Method for Fits}
\description{
Implementation of \code{\link[stats]{predict.glm}} for GLMs fit 
using \code{glm2(..., method = "\link{glm.fit2.Matrix}")}. See its documentation
for full details.
}
