% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glm2.R
\name{vcov.glm2Matrix}
\alias{vcov.glm2Matrix}
\alias{vcov.summary.glm2Matrix}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\description{
Implementation of \code{\link[stats]{vcov}} for GLMs fit 
using \code{glm2(..., method = "\link{glm.fit2.Matrix}")}. See its documentation
for full details.
}
